backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
djangorestframework==3.16.1
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
